chore: add github actions
